requirements.txt
# Core dependencies for mobile device automation
adbutils==0.11.0
uiautomator2==2.16.17
numpy>=1.26.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0

# OCR support
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Utility
retry==0.9.2
retrying==1.3.3
pyyaml>=6.0
logzero==1.7.0
psutil>=5.9.3
lxml>=4.9.0
imageio>=2.27.0
